Extract OrderItem class from bootstrap to a new file
Remove OrderItem class from 'bootrstrap.rb' and create a new file for it
on the project tree (lib/models).

Create test/spec file for OrderItem class.
